third_party/WebKit/LayoutTests/fast/dom/Document/document-open-02.html

<!DOCTYPE HTML>
<title>document.open test for creating new window</title>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<body>
<script>
test(function() {
    window.open = function() { alert('FAIL: The call should be redirected to the real window.open');};
    w = document.open(1, 1, 1);
    assert_equals(typeof(w), 'object', 'w should be a window object');
});
</script>
</body>
